pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gc-memory"
version = "0.1.0"
description = "Self-refining memory store via germinal center dynamics"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.7",
    "sentence-transformers>=2.0",
    "numpy>=1.24",
    "beir>=2.0",
    "datasets>=2.0",
    "matplotlib>=3.7",
    "tqdm>=4.60",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gc_memory"]

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = ["faiss", "datasets", "beir.*", "torch", "torch.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
